Update App test to support providing manually created client secret

+++ SOA/SOA-Prerequisites.psm1
@@ -372,7 +372,7 @@ Function Invoke-AppPermissionCheck
     $SleepTime = 10
     $Counter = 0
 
-    Write-Verbose "$(Get-Date) Invoke-AppPermissionCheck App ID $($App.AppId) Role Count $($Roles.Count)"
+    Write-Verbose "$(Get-Date) Invoke-AppPermissionCheck; App ID $($App.AppId); Role Count $($Roles.Count)"
 
     While($Counter -lt $MaxTime)
     {
@@ -409,11 +409,12 @@ Function Invoke-AppPermissionCheck
 
         If($Provisioned -eq $True)
         {
-            Write-Verbose "$(Get-Date) Invoke-AppPermissionCheck App ID $($app.Id) Role Count $($Roles.Count) OK"
+            Write-Verbose "$(Get-Date) Invoke-AppPermissionCheck; App ID $($app.Id); Role Count $($Roles.Count) OK"
             Break
         } 
         Else 
         {
+            Write-Verbose "$(Get-Date) Invoke-AppPermissionCheck; App ID $($app.Id); Missing roles: $($Missing -Join ";")"
             Start-Sleep $SleepTime
             $Counter += $SleepTime
             Write-Verbose "$(Get-Date) Invoke-AppPermissionCheck loop - waiting for permissions on Entra application - Counter $Counter maxTime $MaxTime Missing $($Missing -join ' ')"
@@ -1714,25 +1715,26 @@ Function Test-SOAApplication
         $Secret,
         $TenantDomain,
         [Switch]$WriteHost,
+        [Switch]$ManualCred=$False,
         [Switch]$NewTokens,
         [Alias("O365EnvironmentName")][string]$CloudEnvironment="Commercial"
     )
 
     Write-Verbose "$(Get-Date) Test-SOAApplication App $($App.AppId) TenantDomain $($TenantDomain) SecretLength $($Secret.Length) CloudEnvironment $CloudEnvironment"
 
-    # Perform permission check
-    If($WriteHost) { Write-Host "$(Get-Date) Performing application permission check... (This may take up to 5 minutes)" }
-    $PermCheck = Invoke-AppPermissionCheck -App $App
+    # Perform permission check, except when manually providing the secret because there will be no delegated connection
+    if ($ManualCred -eq $False) {
+        If($WriteHost) { Write-Host "$(Get-Date) Performing application permission check... (This may take up to 5 minutes)" }
+        $PermCheck = Invoke-AppPermissionCheck -App $App
+    }
 
     # Perform check for consent
-    If($PermCheck -eq $True)
-    {
-        If($WriteHost) { Write-Host "$(Get-Date) Performing token check... (This may take up to 5 minutes)" }
-        If ($NewTokens){
-            $TokenCheck = Invoke-AppTokenRolesCheckV2 -CloudEnvironment $CloudEnvironment
-        } Else {
-            $TokenCheck = Invoke-AppTokenRolesCheck -App $App -Secret $Secret -TenantDomain $tenantdomain -CloudEnvironment $CloudEnvironment
-        }
+    if ($PermCheck -eq $True) {
+        If ($WriteHost) { Write-Host "$(Get-Date) Performing token check... (This may take up to 5 minutes)" }
+        $TokenCheck = Invoke-AppTokenRolesCheckV2 -CloudEnvironment $CloudEnvironment
+    } else {
+        # Set as False to ensure the final result shows the check as Failed instead of Null
+        $TokenCheck = $False
     }
 
     Return New-Object -TypeName PSObject -Property @{
@@ -1782,7 +1784,10 @@ Function Install-SOAPrerequisites
         [Alias('AzureADAppOnly')][switch]$EntraAppOnly,
     [Parameter(ParameterSetName='Default')]
     [Parameter(ParameterSetName='EntraAppOnly')]
-        [switch]$RemoveExistingEntraApp
+        [switch]$RemoveExistingEntraApp,
+    [Parameter(ParameterSetName='Default')]
+    [Parameter(ParameterSetName='EntraAppOnly')]
+        [switch]$PromptForApplicationSecret
     )
 
     <#
@@ -1809,7 +1814,7 @@ Function Install-SOAPrerequisites
     $EntraAppCheck = $True
 
     # Default to remediate (applicable only when not using ConnectOnly)
-    if ($DoNotRemediate -eq $false){
+    if ($DoNotRemediate -eq $false -and $PromptForApplicationSecret -eq $false){
         $Remediate = $true
     }
     else {
@@ -2097,8 +2102,9 @@ Function Install-SOAPrerequisites
             "Germany"      {$cloud = 'Germany'}
             "China"        {$cloud = 'China'}
         }
+
         $mgContext =  (Get-MgContext).Scopes
-        if ($mgContext -notcontains 'Application.ReadWrite.All' -or ($mgContext -notcontains 'Organization.Read.All' -and $mgContext -notcontains 'Directory.Read.All')) {
+        if ($mgContext -notcontains 'Application.ReadWrite.All' -or ($mgContext -notcontains 'Organization.Read.All' -and $mgContext -notcontains 'Directory.Read.All') -or ($PromptForApplicationSecret)) {
             Write-Host "$(Get-Date) Connecting to Graph with delegated authentication..."
             if ($null -ne (Get-MgContext)){Disconnect-MgGraph | Out-Null}
             $connCount = 0
@@ -2107,7 +2113,12 @@ Function Install-SOAPrerequisites
                 try {
                     $connCount++
                     Write-Verbose "$(Get-Date) Graph Delegated connection attempt #$connCount"
-                    Connect-MgGraph -Scopes 'Application.ReadWrite.All','Organization.Read.All' -Environment $cloud -ContextScope "Process" | Out-Null
+                    if ($PromptForApplicationSecret) {
+                        # Request read-only permissions to Graph if manually providing the client secret
+                        Connect-MgGraph -Scopes 'Application.Read.All','Organization.Read.All' -Environment $cloud -ContextScope "Process" | Out-Null
+                    } else {
+                        Connect-MgGraph -Scopes 'Application.ReadWrite.All','Organization.Read.All' -Environment $cloud -ContextScope "Process" | Out-Null
+                    }
                 }
                 catch {
                     Write-Verbose $_
@@ -2135,7 +2146,7 @@ Function Install-SOAPrerequisites
             $script:ATPP2Licensed = Get-LicenseStatus -LicenseType ATPP2
             Write-Verbose "$(Get-Date) Get-LicenseStatus ATPP2 License found: $($script:ATPP2Licensed)"
 
-            # Determine if Microsoft Entra application exists (and has public client redirect URI set), create if doesnt
+            # Determine if Microsoft Entra application exists (and has public client redirect URI set), create if doesn't
             $EntraApp = Get-SOAEntraApp -CloudEnvironment $CloudEnvironment
         }
 
@@ -2157,14 +2168,26 @@ Function Install-SOAPrerequisites
                 }
             }
 
-            # Reset secret
-            $clientsecret = Reset-SOAAppSecret -App $EntraApp -Task "Prereq"
-            Write-Host "$(Get-Date) Sleeping to allow for replication of the application's new client secret..."
-            Start-Sleep 10
+            if ($PromptForApplicationSecret -eq $True) {
+                # Prompt for the client secret needed to connect to the application
+                $SSCred = $null
+
+                Write-Host "$(Get-Date) At the prompt, provide a valid client secret for the assessment's app registration."
+                Start-Sleep -Seconds 1
+                while ($null -eq $SSCred -or $SSCred.Length -eq 0) {
+                    # UserName is a required parameter for Get-Credential but it's value is not used elsewhere in the script
+                    $SSCred = (Get-Credential -Message "Enter the app registration's client secret into the password field." -UserName "Microsoft Security Assessment").Password
+                }
+            } else {
+                # Reset secret
+                $clientsecret = Reset-SOAAppSecret -App $EntraApp -Task "Prereq"
+                $SSCred = $clientsecret | ConvertTo-SecureString -AsPlainText -Force
+                Write-Host "$(Get-Date) Sleeping to allow for replication of the app registration's new client secret..."
+                Start-Sleep 10
+            }
 
             # Reconnect with Application permissions
             Disconnect-MgGraph | Out-Null
-            $SSCred = $clientsecret | ConvertTo-SecureString -AsPlainText -Force
             $GraphCred = New-Object -TypeName System.Management.Automation.PSCredential -ArgumentList ($EntraApp.AppId), $SSCred
             $ConnCount = 0
             Write-Host "$(Get-Date) Connecting to Graph with application authentication..."
@@ -2181,7 +2204,7 @@ Function Install-SOAPrerequisites
             $AppTest = Test-SOAApplication -App $EntraApp -Secret $clientsecret -TenantDomain $tenantdomain -CloudEnvironment $CloudEnvironment -WriteHost
                 
             # Entra App Permission - Perform remediation if specified
-            If($AppTest.Permissions -eq $False -and $DoNotRemediate -eq $false)
+            If($AppTest.Permissions -eq $False -and $Remediate -eq $true)
             {
                 # Set up the correct Entra App Permissions
                 Write-Host "$(Get-Date) Remediating application permissions..."
@@ -2208,7 +2231,7 @@ Function Install-SOAPrerequisites
             If($AppTest.Token -eq $False)
             {
                 Write-Host "$(Get-Date) Missing roles in access token; possible that consent was not completed..."
-                if ($DoNotRemediate -eq $false) {
+                if ($Remediate -eq $true) {
                     # Request admin consent
                     If((Invoke-Consent -App $EntraApp -CloudEnvironment $CloudEnvironment) -eq $True) {
                         # Perform check again after consent
@@ -2245,8 +2268,10 @@ Function Install-SOAPrerequisites
                     Pass=$True
                 }
 
-                # Remove client secret
-                Remove-SOAAppSecret
+                if ($PromptForApplicationSecret -eq $false) {
+                    # Remove client secret
+                    Remove-SOAAppSecret
+                }
                 # Disconnect
                 Disconnect-MgGraph | Out-Null
             }
